type cast too strong corrections

+++ src/main/java/net/perspective/draw/geom/Figure.java
@@ -538,7 +538,7 @@ public class Figure implements DrawItem, Serializable {
     }
 
     public double getLineWidth(BasicStroke stroke) {
-        return (double) (stroke.getLineWidth());
+        return stroke.getLineWidth();
     }
     
     public StrokeLineJoin getLineJoin(BasicStroke stroke) {
